Tooltips
Tooltips can now be optionally provided for the bar and line charts. The tooltips are rendered as <title> elements within these lines.

Scatter charts are not catered for as there is no svg container element such as a line. However, given that scatter charts are rendered using labels, the labels themselves can include their own <title> element if required.